+++ example/main.dart
@@ -9,11 +9,16 @@ Future<void> main(List<String> arguments) async {
 
   //one way of manually importing keys
   keyz = WebPushKeys(
-    pubKeyBytes: base64.decode(
-        "BGviCUiE9bL6HqxXZRLKb3pmHYGq24acYDoE-Hy2aZM9h2gIx0jrQTWh2ksIaFegv6yUQLkpbV7984w0IpvlT-Y="),
-    authKeyBytes: Uint8List(16),
-    privKeyBytes: base64Decode(
-        "MIGHAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBG0wawIBAQQga-gpVIxZJHJ_uUx9ON45Lm5Owp5qbM3o7u0p0qrmoZqhRANCAARr4glIhPWy-h6sV2USym96Zh2BqtuGnGA6BPh8tmmTPYdoCMdI60E1odpLCGhXoL-slEC5KW1e_fOMNCKb5U_m"),
+    publicKey: WebPublicPushKey(
+      publicKeyBytes: base64.decode(
+          "BGviCUiE9bL6HqxXZRLKb3pmHYGq24acYDoE-Hy2aZM9h2gIx0jrQTWh2ksIaFegv6yUQLkpbV7984w0IpvlT-Y="),
+      authKeyBytes: Uint8List(16),
+    ),
+    privateKey: WebPrivatePushKey(
+      privateKeyBytes: base64Decode(
+        "MIGHAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBG0wawIBAQQga-gpVIxZJHJ_uUx9ON45Lm5Owp5qbM3o7u0p0qrmoZqhRANCAARr4glIhPWy-h6sV2USym96Zh2BqtuGnGA6BPh8tmmTPYdoCMdI60E1odpLCGhXoL-slEC5KW1e_fOMNCKb5U_m",
+      ),
+    ),
   );
   // this is equivalent to the above key
   assert(keyz ==
@@ -21,8 +26,8 @@ Future<void> main(List<String> arguments) async {
           "BGviCUiE9bL6HqxXZRLKb3pmHYGq24acYDoE-Hy2aZM9h2gIx0jrQTWh2ksIaFegv6yUQLkpbV7984w0IpvlT-Y=+AAAAAAAAAAAAAAAAAAAAAA==+MIGHAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBG0wawIBAQQga-gpVIxZJHJ_uUx9ON45Lm5Owp5qbM3o7u0p0qrmoZqhRANCAARr4glIhPWy-h6sV2USym96Zh2BqtuGnGA6BPh8tmmTPYdoCMdI60E1odpLCGhXoL-slEC5KW1e_fOMNCKb5U_m")); // importing based on storage format
 
 // this is what the WebPush Application Server needs
-  print(keyz.p256dh);
-  print(keyz.auth);
+  print(keyz.publicKey!.p256dh);
+  print(keyz.publicKey!.auth);
 
 //these two are equivalent except for the format
   print(keyz.serialize);