+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.0
+
+- WebPush encryption and decryption.
+- Breaking changes:
+   - WebPublicPushKey - new class
+   - WebPrivatePushKey - new class
+   - WebPushKeys - overhauled as featured KeyPair
+
 ## 0.1.0
 
 - Initial release